@@ -429,29 +428,29 @@
         "tsconfig-paths-webpack-plugin": "^3.5.1",
         "tslint": "~6.1.0",
         "ttypescript": "^1.5.12",
-        "typescript": "~4.4.2",
+        "typescript": "~4.4.3",
         "typescript-plugin-css-modules": "^3.4.0",
         "url-loader": "^4.1.1",
-        "vite": "^2.5.3",
+        "vite": "^2.5.10",
         "watchpack": "^2.2.0",
-        "webpack": "^5.51.1",
+        "webpack": "^5.53.0",
         "webpack-bundle-analyzer": "^4.4.2",
         "webpack-chokidar-plugin": "^1.0.5",
         "webpack-cli": "^4.8.0",
         "webpack-copy-after-build-plugin": "^1.0.1",
-        "webpack-dev-middleware": "^5.0.0",
-        "webpack-dev-server": "^4.1.0",
+        "webpack-dev-middleware": "^5.1.0",
+        "webpack-dev-server": "^4.2.1",
         "webpack-fix-style-only-entries": "^0.6.1",
         "webpack-glob-entry": "^2.1.1",
-        "webpack-hot-middleware": "^2.25.0",
+        "webpack-hot-middleware": "^2.25.1",
         "webpack-manifest-plugin": "^4.0.2",
         "webpack-merge": "^5.8.0",
         "webpack-pwa-manifest": "^4.3.0",
-        "webpack-sources": "^3.2.0",
+        "webpack-sources": "^3.2.1",
         "webpack-watch-files-plugin": "^1.1.0",
         "webpack-watched-glob-entries-plugin": "^2.2.2",
-        "workbox-build": "^6.2.4",
-        "workbox-webpack-plugin": "^6.2.4",
+        "workbox-build": "^6.3.0",
+        "workbox-webpack-plugin": "^6.3.0",
         "xml-loader": "^1.2.1"
     },
     "peerDependencies": {
