@@ -3,7 +3,7 @@ import tailwindcssElevation from 'tailwindcss-elevation';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 // const plugin = require('tailwindcss/plugin');
-const plugin = require('tailwindcss/plugin')
+import plugin from 'tailwindcss/plugin';
 import { materialVariables } from './src/web/material/variables/export';
 import { tailwindColors } from './src/web/tailwind/tailwind.colors';
 import tailwindImportant from 'tailwindcss-important';
