@@ -1,18 +1,18 @@
-import { EnvironmentService } from '../../../modules/environment/environment/environment.service';
-import { CssnanoService } from '../services/cssnano/cssnano.service';
-import { PostcssFontMagicianService } from '../services/postcss-font-magician/postcss-font-magician.service';
-import { PostcssFunctionsService } from '../services/postcss-functions/postcss-functions.service';
-import { PostcssPresetEnvService } from '../services/postcss-preset-env/postcss-preset-env.service';
-import { PostcssReporterService } from '../services/postcss-reporter/postcss-reporter.service';
-import { PostcssRfsAutopilotService } from '../services/postcss-rfs-autopilot/postcss-rfs-autopilot.service';
-import { RfsService } from '../services/rfs/rfs.service';
-import { TailwindcssService } from '../services/tailwindcss/tailwindcss.service';
+import { EnvironmentService } from './../../modules/environment/environment/environment.service';
+import { CssnanoService } from './services/cssnano/cssnano.service';
+import { PostcssFontMagicianService } from './services/postcss-font-magician/postcss-font-magician.service';
+import { PostcssFunctionsService } from './services/postcss-functions/postcss-functions.service';
+import { PostcssPresetEnvService } from './services/postcss-preset-env/postcss-preset-env.service';
+import { PostcssReporterService } from './services/postcss-reporter/postcss-reporter.service';
+import { PostcssRfsAutopilotService } from './services/postcss-rfs-autopilot/postcss-rfs-autopilot.service';
+import { RfsService } from './services/rfs/rfs.service';
+// import { TailwindcssService } from '../services/tailwindcss/tailwindcss.service';
 
 // Sealed class.
 import { CustomInjectable } from '#shared/src/functions/process-providers';
 import { CustomInject } from '#shared/src/functions/process-providers';
-import { PostcssPurgecssService } from '../services/postcss-purgecss/postcss-purgecss.service';
-import { PostcssCombineDuplicatedSelectorsService } from '../services/postcss-combine-duplicated-selectors/postcss-combine-duplicated-selectors.service';
+import { PostcssPurgecssService } from './services/postcss-purgecss/postcss-purgecss.service';
+import { PostcssCombineDuplicatedSelectorsService } from './services/postcss-combine-duplicated-selectors/postcss-combine-duplicated-selectors.service';
 
 @CustomInjectable()
 export class PostcssService {
@@ -32,8 +32,8 @@ export class PostcssService {
     protected PostcssFunctionsService: PostcssFunctionsService;
     @CustomInject(PostcssFontMagicianService)
     protected PostcssFontMagicianService: PostcssFontMagicianService;
-    @CustomInject(TailwindcssService)
-    protected TailwindcssService: TailwindcssService;
+    // @CustomInject(TailwindcssService)
+    // protected TailwindcssService: TailwindcssService;
     @CustomInject(PostcssRfsAutopilotService)
     protected PostcssRfsAutopilotService: PostcssRfsAutopilotService;
     @CustomInject(RfsService)
@@ -43,7 +43,7 @@ export class PostcssService {
         plugins.push(this.PostcssRfsAutopilotService.createPlugin());
         plugins.push(this.RfsService.createPlugin());
 
-        plugins.push(this.TailwindcssService.createPlugin());
+        // plugins.push(this.TailwindcssService.createPlugin());
         // plugins.push(this.PostcssFunctionsService.createPlugin());
         // plugins.push(this.PostcssFontMagicianService.createPlugin());
         // plugins.push(this.PostcssPresetEnvService.createPlugin());
