@@ -1,7 +1,7 @@
 import { ConfigBase } from '#shared/configs.base';
 import { solidAppPaths } from './paths';
 
-export class BlazorAppConfig extends ConfigBase {
+export class SolidAppConfig extends ConfigBase {
     rootDir = solidAppPaths.toAbsolutePath();
     RootDir = solidAppPaths.toAbsolutePath();
     projectName = 'solidrApp';
@@ -13,4 +13,4 @@ export class BlazorAppConfig extends ConfigBase {
     description = 'app desc';
 }
 
-export const blazorAppConfig = new BlazorAppConfig();
+export const solidAppConfig = new SolidAppConfig();
