@@ -2,4 +2,4 @@ import { CustomInjectable } from '#shared/src/functions/process-providers';
 import { EnvironmentService } from '#shared/src/modules/environment/environment/environment.service';
 
 @CustomInjectable()
-export class BlazorAppEnvironmentService extends EnvironmentService {}
+export class SolidAppEnvironmentService extends EnvironmentService {}
