@@ -1,11 +1,11 @@
 import { CustomModule } from '#shared/src/functions/process-providers';
 import { PostcssModule } from '#shared/src/modules/postcss/postcss.module';
 import { Module } from '@nestjs/common';
-import { BlazorAppPostcssService } from './postcss/postcss.service';
+import { SolidAppPostcssService } from './postcss/postcss.service';
 
 
 @CustomModule({
     imports: [PostcssModule],
-    providers: [BlazorAppPostcssService],
+    providers: [SolidAppPostcssService],
 })
 export class BlazorWebPostcssModule {}
