@@ -1,3 +1,3 @@
 import { MergeCommand } from '#shared/src/modules/utilities/modules/merge/merge/merge-command';
 
-export class BlazorAppMergeCommand extends MergeCommand {}
+export class SolidAppMergeCommand extends MergeCommand {}
