@@ -3,4 +3,4 @@ import { RegexService } from '#shared/src/modules/utilities/modules/regex/regex/
 import { CustomInjectable } from '#shared/src/functions/process-providers';
 
 @CustomInjectable()
-export class BlazorAppRegexService extends RegexService {}
+export class SolidAppRegexService extends RegexService {}
