@@ -1,10 +1,10 @@
 import { CustomModule } from '#shared/src/functions/process-providers';
-import { MergeModule as BlazorAppMergeModule } from './merge/merge.module';
+import { MergeModule as SolidAppMergeModule } from './merge/merge.module';
 
-import { RegexModule as BlazorAppRegexModule } from './regex/regex.module';
+import { RegexModule as SolidAppRegexModule } from './regex/regex.module';
 
 @CustomModule({
-    imports: [BlazorAppMergeModule, BlazorAppRegexModule],
+    imports: [SolidAppMergeModule, SolidAppRegexModule],
     providers: [],
 })
-export class BlazorAppUtilitiesModule {}
+export class SolidAppUtilitiesModule {}
