@@ -1,11 +1,11 @@
 import { CustomModule } from "#shared/src/functions/process-providers";
-import { BlazorAppVitePluginsService } from "./vite-plugins.service";
+import { SolidAppVitePluginsService } from "./vite-plugins.service";
 
 
 @CustomModule({
     imports: [],
     providers: [
-        BlazorAppVitePluginsService
+        SolidAppVitePluginsService
     ],
 })
-export class BlazorAppVitePluginsModule {}
+export class SolidAppVitePluginsModule {}
