@@ -6,23 +6,23 @@ import { ViteBaseService } from '#shared/src/vite/vite-base/vite-base.service';
 import { UserConfig } from 'vite';
 import path from 'path';
 
-import { BlazorAppEnvironmentService } from '#solid-app/src/modules/environment/environment/environment.service';
+import { SolidAppEnvironmentService } from '#solid-app/src/modules/environment/environment/environment.service';
 import { ViteSharedService } from '#shared/src/vite/vite-shared/vite-shared.service';
-import { BlazorAppVitePluginsService } from '../vite-plugins/vite-plugins.service';
+import { SolidAppVitePluginsService } from '../vite-plugins/vite-plugins.service';
 import { RootPaths, rootPaths } from '#root/paths';
-import { PostcssService } from '#shared/src/modules/postcss/postcss/postcss.service';
-import { BlazorAppPostcssService } from '#solid-app/src/modules/postcss/postcss/postcss.service';
+import { PostcssService } from '#shared/src/modules/postcss/postcss.service';
+import { SolidAppPostcssService } from '#solid-app/src/modules/postcss/postcss/postcss.service';
 @CustomInjectable()
-export class BlazorAppViteSharedService extends ViteSharedService {
-    @CustomInject(BlazorAppEnvironmentService)
-    protected blazorAppEnvironmentService: BlazorAppEnvironmentService;
-    @CustomInject(BlazorAppVitePluginsService)
-    protected blazorAppVitePluginsService: BlazorAppVitePluginsService;
-    @CustomInject(BlazorAppPostcssService)
-    protected blazorAppPostcssService: BlazorAppPostcssService;
+export class SolidAppViteSharedService extends ViteSharedService {
+    @CustomInject(SolidAppEnvironmentService)
+    protected solidAppEnvironmentService: SolidAppEnvironmentService;
+    @CustomInject(SolidAppVitePluginsService)
+    protected solidAppVitePluginsService: SolidAppVitePluginsService;
+    @CustomInject(SolidAppPostcssService)
+    protected solidAppPostcssService: SolidAppPostcssService;
     createConfiguration(options?: UserConfig) {
-        const postcssPlugins = this.blazorAppPostcssService.createPostcssPlugins();
-        const plugins = this.blazorAppVitePluginsService.createManyPlugins();
+        const postcssPlugins = this.solidAppPostcssService.createPostcssPlugins();
+        const plugins = this.solidAppVitePluginsService.createManyPlugins();
         return this.mergeService.mergeOptions(
             super.createConfiguration(), {
 
