@@ -1,10 +1,10 @@
 
 
 import { ViteBase } from '#shared/vite.base';
-import { BlazorAppViteDevService } from './src/vite/vite-dev/vite-dev.service';
-import { BlazorAppViteModule } from './src/vite/vite.module';
+import { SolidAppViteDevService } from './src/vite/vite-dev/vite-dev.service';
+import { SolidAppViteModule } from './src/vite/vite.module';
 
-const viteBase = new ViteBase(BlazorAppViteModule, BlazorAppViteDevService);
+const viteBase = new ViteBase(SolidAppViteModule, SolidAppViteDevService);
 
 await viteBase.build();
 
