@@ -1,10 +1,10 @@
 
 
 import { ViteBase } from '#shared/vite.base';
-import { BlazorAppViteProdService } from './src/vite/vite-prod/vite-prod.service';
-import { BlazorAppViteModule } from './src/vite/vite.module';
+import { SolidAppViteProdService } from './src/vite/vite-prod/vite-prod.service';
+import { SolidAppViteModule } from './src/vite/vite.module';
 
-const viteBase = new ViteBase(BlazorAppViteModule, BlazorAppViteProdService);
+const viteBase = new ViteBase(SolidAppViteModule, SolidAppViteProdService);
 
 viteBase.build();
 
