@@ -1,6 +1,5 @@
-const postcss = require('postcss');
 import { shouldBeTransformed } from './utilities/helper';
-export default postcss.plugin('postcss-rfs-autopilot', ({ includedRules, excludedRules, includedSelectors, excludedSelectors, includedUnits, excludedUnits, silentConsole }) => {
+const plugin = (({ includedRules, excludedRules, includedSelectors, excludedSelectors, includedUnits, excludedUnits, silentConsole }) => {
     const options = {
         includedRules: includedRules || [
             '*'
@@ -18,12 +17,17 @@ export default postcss.plugin('postcss-rfs-autopilot', ({ includedRules, exclude
     // options.includedRules = filterIdenticalValues(options.includedRules, options.excludedRules)
     // options.includedSelectors = filterIdenticalValues(options.includedSelectors, options.excludedSelectors)
     // options.includedUnits = filterIdenticalValues(options.includedUnits, options.excludedUnits)
-    return (root, result) => {
-        root.walkDecls((decl) => {
-            if (shouldBeTransformed(decl, options)) {
-                decl.value = `rfs(${decl.value})`;
-            }
-        });
+    return {
+        postcssPlugin: 'postcss-rfs-autopilot',
+        Once(root, { result }) {
+            root.walkDecls((decl) => {
+                if (shouldBeTransformed(decl, options)) {
+                    decl.value = `rfs(${decl.value})`;
+                }
+            });
+        }
     };
 });
+plugin['postcss'] = true;
+export default plugin;
 //# sourceMappingURL=index.js.map
