@@ -19,22 +19,7 @@ let BlazorAppVitePluginsService = class BlazorAppVitePluginsService extends Vite
     createManyPlugins() {
         return [
             this.vitePluginHtmlService.createPlugin({
-                externals: [
-                    { html: '<link href="_content/Material.Blazor/Material.min.css" rel="stylesheet" />', insertAt: '<head>' },
-                    { html: '<link href="_content/Material.Blazor/MaterialBlazor.min.css" rel="stylesheet" />', insertAt: '<head>' },
-                    { html: '<link href="BlazorApp.Client.styles.css" rel="stylesheet" />', insertAt: '</head>' },
-                    { html: '<script src="_content/Material.Blazor/MaterialBlazor.min.js"></script>', insertAt: '</body>' },
-                    {
-                        html: `
-                        <script>
-                            MaterialBlazor.MBTooltip.numbers.HIDE_DELAY_MS = 0
-                        </script>
-                        `, insertAt: '</body>'
-                    },
-                    { html: '<script src="_content/Microsoft.AspNetCore.Components.WebAssembly.Authentication/AuthenticationService.js"></script>', insertAt: '</body>' },
-                    { html: '<script src="_framework/blazor.webassembly.js"></script>', insertAt: '</body>' },
-                    { html: '<script>navigator.serviceWorker.register("service-worker.js");</script>', insertAt: '</body>' },
-                ]
+                externals: []
             }),
             this.vitePluginGlobInputService.createPlugin({
                 inputs: [
