@@ -8771,10 +8771,6 @@ export const RootPaths = createProxy({
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\appsettings.json",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\auth\\src\\appsettings.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\auth\\src\\appsettings.json"); }
                     },
-                    "index.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\index.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\auth\\src\\index.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\auth\\src\\index.ts"); }
-                    },
                     "logo": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\logo",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\auth\\src\\logo") : path.relative("C:\\App", "C:\\App\\Js\\apps\\auth\\src\\logo"); },
@@ -8790,10 +8786,6 @@ export const RootPaths = createProxy({
                     "modules": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\modules",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\auth\\src\\modules") : path.relative("C:\\App", "C:\\App\\Js\\apps\\auth\\src\\modules"); }
-                    },
-                    "server.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\server.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\auth\\src\\server.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\auth\\src\\server.ts"); }
                     },
                     "service-worker": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\auth\\src\\service-worker",
@@ -8963,10 +8955,6 @@ export const RootPaths = createProxy({
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\appsettings.json",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\blazor-app\\src\\appsettings.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\blazor-app\\src\\appsettings.json"); }
                     },
-                    "index.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\index.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\blazor-app\\src\\index.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\blazor-app\\src\\index.ts"); }
-                    },
                     "logo": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\logo",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\blazor-app\\src\\logo") : path.relative("C:\\App", "C:\\App\\Js\\apps\\blazor-app\\src\\logo"); },
@@ -8982,10 +8970,6 @@ export const RootPaths = createProxy({
                     "modules": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\modules",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\blazor-app\\src\\modules") : path.relative("C:\\App", "C:\\App\\Js\\apps\\blazor-app\\src\\modules"); }
-                    },
-                    "server.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\server.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\blazor-app\\src\\server.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\blazor-app\\src\\server.ts"); }
                     },
                     "service-worker": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\blazor-app\\src\\service-worker",
@@ -9271,10 +9255,6 @@ export const RootPaths = createProxy({
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\modules",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\modules") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\modules"); }
                     },
-                    "server.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\server.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\server.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\server.ts"); }
-                    },
                     "service-worker": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\service-worker",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\service-worker") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\service-worker"); },
@@ -9329,6 +9309,34 @@ export const RootPaths = createProxy({
                         "vite-plugins": {
                             toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins",
                             toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins"); },
+                            "services": {
+                                toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services",
+                                toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services"); },
+                                "vite-plugin-glob-input": {
+                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input",
+                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input"); },
+                                    "vite-plugin-glob-input.service.spec.ts": {
+                                        toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts",
+                                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts"); }
+                                    },
+                                    "vite-plugin-glob-input.service.ts": {
+                                        toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts",
+                                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts"); }
+                                    }
+                                },
+                                "vite-plugin-html": {
+                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html",
+                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html"); },
+                                    "vite-plugin-html.service.spec.ts": {
+                                        toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.spec.ts",
+                                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.spec.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.spec.ts"); }
+                                    },
+                                    "vite-plugin-html.service.ts": {
+                                        toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.ts",
+                                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\services\\vite-plugin-html\\vite-plugin-html.service.ts"); }
+                                    }
+                                }
+                            },
                             "vite-plugin-base": {
                                 toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base",
                                 toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base"); },
@@ -9339,30 +9347,6 @@ export const RootPaths = createProxy({
                                 "vite-plugin-base.service.ts": {
                                     toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base\\vite-plugin-base.service.ts",
                                     toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base\\vite-plugin-base.service.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-base\\vite-plugin-base.service.ts"); }
-                                }
-                            },
-                            "vite-plugin-glob-input": {
-                                toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input",
-                                toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input"); },
-                                "vite-plugin-glob-input.service.spec.ts": {
-                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts",
-                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.spec.ts"); }
-                                },
-                                "vite-plugin-glob-input.service.ts": {
-                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts",
-                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-glob-input\\vite-plugin-glob-input.service.ts"); }
-                                }
-                            },
-                            "vite-plugin-html": {
-                                toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html",
-                                toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html"); },
-                                "vite-plugin-html.service.spec.ts": {
-                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.spec.ts",
-                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.spec.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.spec.ts"); }
-                                },
-                                "vite-plugin-html.service.ts": {
-                                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.ts",
-                                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\src\\vite\\vite-plugins\\vite-plugin-html\\vite-plugin-html.service.ts"); }
                                 }
                             },
                             "vite-plugins": {
@@ -10880,18 +10864,6 @@ export const RootPaths = createProxy({
                     toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\tailwind.config.ts",
                     toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\tailwind.config.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\tailwind.config.ts"); }
                 },
-                "test.ts": {
-                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\test.ts",
-                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\test.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\test.ts"); }
-                },
-                "test2.ts": {
-                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\test2.ts",
-                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\test2.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\test2.ts"); }
-                },
-                "tsconfig copy.ts": {
-                    toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\tsconfig copy.ts",
-                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\tsconfig copy.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\tsconfig copy.ts"); }
-                },
                 "tsconfig.backup.json": {
                     toAbsolutePath: () => "C:\\App\\Js\\apps\\shared\\tsconfig.backup.json",
                     toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\shared\\tsconfig.backup.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\shared\\tsconfig.backup.json"); }
@@ -10966,6 +10938,74 @@ export const RootPaths = createProxy({
                     "tasks.json": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\.vscode\\tasks.json",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\.vscode\\tasks.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\.vscode\\tasks.json"); }
+                    }
+                },
+                "app": {
+                    toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app",
+                    toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app"); },
+                    ".gitignore": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\.gitignore",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\.gitignore") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\.gitignore"); }
+                    },
+                    "index.html": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\index.html",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\index.html") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\index.html"); }
+                    },
+                    "node_modules": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\node_modules",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\node_modules") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\node_modules"); }
+                    },
+                    "package-lock.json": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\package-lock.json",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\package-lock.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\package-lock.json"); }
+                    },
+                    "package.json": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\package.json",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\package.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\package.json"); }
+                    },
+                    "pnpm-lock.yaml": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\pnpm-lock.yaml",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\pnpm-lock.yaml") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\pnpm-lock.yaml"); }
+                    },
+                    "README.md": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\README.md",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\README.md") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\README.md"); }
+                    },
+                    "src": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src"); },
+                        "App.tsx": {
+                            toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\App.tsx",
+                            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\App.tsx") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\App.tsx"); }
+                        },
+                        "assets": {
+                            toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets",
+                            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets"); },
+                            "favicon.ico": {
+                                toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets\\favicon.ico",
+                                toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets\\favicon.ico") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\assets\\favicon.ico"); }
+                            }
+                        },
+                        "index.scss": {
+                            toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.scss",
+                            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.scss") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.scss"); }
+                        },
+                        "index.tsx": {
+                            toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.tsx",
+                            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.tsx") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\index.tsx"); }
+                        },
+                        "logo.svg": {
+                            toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\src\\logo.svg",
+                            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\src\\logo.svg") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\src\\logo.svg"); }
+                        }
+                    },
+                    "tsconfig.json": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\tsconfig.json",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\tsconfig.json") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\tsconfig.json"); }
+                    },
+                    "vite.config.ts": {
+                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\app\\vite.config.ts",
+                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\app\\vite.config.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\app\\vite.config.ts"); }
                     }
                 },
                 "berry.json": {
@@ -11046,10 +11086,6 @@ export const RootPaths = createProxy({
                     "modules": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\src\\modules",
                         toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\src\\modules") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\src\\modules"); }
-                    },
-                    "server.ts": {
-                        toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\src\\server.ts",
-                        toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\apps\\solid-app\\src\\server.ts") : path.relative("C:\\App", "C:\\App\\Js\\apps\\solid-app\\src\\server.ts"); }
                     },
                     "service-worker": {
                         toAbsolutePath: () => "C:\\App\\Js\\apps\\solid-app\\src\\service-worker",
@@ -11297,18 +11333,6 @@ export const RootPaths = createProxy({
         "tailwind.config.ts": {
             toAbsolutePath: () => "C:\\App\\Js\\tailwind.config.ts",
             toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\tailwind.config.ts") : path.relative("C:\\App", "C:\\App\\Js\\tailwind.config.ts"); }
-        },
-        "test.ts": {
-            toAbsolutePath: () => "C:\\App\\Js\\test.ts",
-            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\test.ts") : path.relative("C:\\App", "C:\\App\\Js\\test.ts"); }
-        },
-        "test2.js": {
-            toAbsolutePath: () => "C:\\App\\Js\\test2.js",
-            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\test2.js") : path.relative("C:\\App", "C:\\App\\Js\\test2.js"); }
-        },
-        "test2.ts": {
-            toAbsolutePath: () => "C:\\App\\Js\\test2.ts",
-            toRelativePath: (relativeTo) => { return relativeTo ? path.relative(relativeTo, "C:\\App\\Js\\test2.ts") : path.relative("C:\\App", "C:\\App\\Js\\test2.ts"); }
         },
         "tsconfig.json": {
             toAbsolutePath: () => "C:\\App\\Js\\tsconfig.json",
